# deliverables/openapi.yaml
openapi: 3.0.3
info:
  title: Venus Bijoux API
  version: 1.0.0
  description: API para e commerce de bijuteria artesanal com configurador de montagem
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:4000
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        provider:
          type: string
          enum: [email, google, facebook]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
        refreshToken:
          type: string
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
          enum: [brinco, colar, pulseira]
        hasPendant:
          type: boolean
        images:
          type: array
          items:
            type: string
    ProductList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
paths:
  /api/auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/refresh:
    post:
      summary: Refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Tokens renewed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
  /api/products:
    get:
      summary: List products with filters and pagination
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: A list of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      summary: Create product (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "201":
          description: Created
  /api/products/{slug}:
    get:
      summary: Get product detail by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      summary: Update product (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete product (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /api/pendants:
    get:
      summary: List pendant options (for earrings)
      responses:
        "200":
          description: A list of pendant options
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    color:
                      type: string
                      enum: [dourado, prata, bronze]
                    priceAdjustment:
                      type: number
  /api/cart:
    get:
      summary: Get current cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart
    post:
      summary: Add item to cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productVariantId:
                  type: string
                pendantOptionId:
                  type: string
                quantity:
                  type: integer
      responses:
        "200":
          description: Item added
  /api/checkout:
    post:
      summary: Create checkout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                shippingAddress:
                  type: object
                paymentMethod:
                  type: string
      responses:
        "201":
          description: Checkout created with payment link
  /api/orders:
    get:
      summary: List orders for user or admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orders list
  /api/uploads:
    post:
      summary: Request upload url (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                contentType:
                  type: string
      responses:
        "200":
          description: Presigned url
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadUrl:
                    type: string
                  publicUrl:
                    type: string
  /api/webhooks/payments:
    post:
      summary: Payment provider webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: handled
security:
  - bearerAuth: []